be/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
motor
pydantic
boto3
gunicorn
python-multipart